requirements.txt
Django==3.2.5
python-decouple==3.4
dj-database-url==0.5.0
mysqlclient==2.0.3
psycopg2-binary==2.9.1
djangorestframework==3.12.4
markdown==3.3.4
